requirements.txt
faiss==1.8.0
fastapi==0.115.6
jinja2==3.1.5
langchain_community==0.2.19
langchain_core==0.2.43
langchain_openai==0.1.25
numpy==1.24.4
pandas==2.0.3
pydantic==2.10.5
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
uvicorn==0.33.0
